requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.5
redis==4.5.4
django-redis==5.2.0
pytest==7.3.1
pytest-django==4.5.2
djangorestframework-simplejwt==5.2.2
bcrypt==4.0.1
gunicorn==20.1.0
dj-database-url==1.3.0
django-cors-headers==3.14.0
django-password-validators==1.7.0
drf-yasg==1.21.5
django-environ==0.10.0